Fix: don't change scroll mode on layer changes
The keyboard-level layer_state_set_kb function for aki27/cocot36plus_v2
was changing the scroll mode depending on the layer number, and also
failing to call the user-level function.

This makes unwarranted assumptions about the user's layers, and also did
the wrong thing when an automatic mouse layer was being used.

Removed the layer_state_set_kb function entirely. The things that it was
trying to do are better done at the keymap level.