Import weteor/qazikat from weteor/Qazikat
Update the matrix, which had one too many columns for Qazikat (as
opposed to Katiqaz), and the layout, which was apparently for a
different keyboard entirely.

Define a separate named layout for each of the possible bottom rows, and
update the default keymap to use the fully split space layout.