Fix cocot38mini config struct
It was incorrectly trying to use a 64-bit structure, when
eeconfig_update_kb in fact takes a uint32_t as its argument. This meant
that values after the first 32 bits were being ignored, and since it was
not packing its bool field into single bits, that was everything after
`auto_mouse` (in particular, `scrl_inv` was not stored).

Removed the `last_mouse` field entirely, since that was not being used
(and would not fit).